Fix signed/unsigned warnings

+++ Source/SIMPLib/Geometry/EdgeGeom.cpp
@@ -80,7 +80,7 @@ class FindEdgeDerivativesImpl
       double derivs[3] = { 0.0, 0.0, 0.0 };
       int64_t verts[2] = { 0, 0 };
 
-      size_t counter = 0;
+      int64_t counter = 0;
       int64_t totalElements = m_Edges->getNumberOfEdges();
       int64_t progIncrement = static_cast<int64_t>(totalElements / 100);
