+++ Source/SIMPLib/Geometry/QuadGeom.cpp
@@ -80,7 +80,7 @@ class FindQuadDerivativesImpl
       double derivs[3] = { 0.0, 0.0, 0.0 };
       int64_t verts[4] = { 0, 0, 0, 0 };
 
-      size_t counter = 0;
+      int64_t counter = 0;
       int64_t totalElements = m_Quads->getNumberOfQuads();
       int64_t progIncrement = static_cast<int64_t>(totalElements / 100);
